[ADD] Added a service to work with session table

+++ backend/src/user/services/avatar.service.ts
@@ -37,7 +37,10 @@ export class AvatarService {
 
   /* DELETE */
 
-  async deleteAvatar(id: number): Promise<void> {
+  async delete(id: number): Promise<void> {
+    if (!user) throw new HttpException('Body is null', HttpStatus.NOT_FOUND);
+    await this.findById(id);
+
     try {
       await this.avatarRepo.delete(id);
     } catch (error) {
